% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{betwixt}
\alias{betwixt}
\title{Find values between lower and upper bounds of a range}
\usage{
betwixt(x, lower, upper, ignore = NULL, by = 1, value = FALSE)
}
\arguments{
\item{x}{Vector of numeric or date values}

\item{lower}{The lower bound (inclusive) for the range of values}

\item{upper}{The upper bound (inclusive) for the range of values}

\item{ignore}{Vector containing value(s) to exclude; default is \code{NULL}}

\item{by}{Value for increment between lower and upper bounds of the range; default is \code{1}}

\item{value}{Boolean as to whether the function should return values of \code{x} between bounds (\code{TRUE}) or a vector of logical indices corresponding to whether or not the value falls between bounds (\code{FALSE}); default is \code{FALSE}}
}
\value{
Depending on the "value" argument, this will either return a vector of values that fall between the lower and upper bounds \emph{or} a vector of logical vector with \code{TRUE} for each index corresponding to a value that falls in the range.
}
\description{
This is utility function to find values of a vector tha fall between lower and upper bounds specified. The implementation includes a method to optionally ignore certain values and return either a  vector of logical indices or a vector of matching values. Inspired by \link[dplyr]{between}.
}
\examples{

betwixt(1:12, 7, 9)
betwixt(1:12, 7, 9, value = TRUE)

dates <- seq(as.Date("2000/1/1"), by = "day", length.out = 365)
betwixt(dates,
        lower = as.Date("2000/1/10"),
        upper = as.Date("2000/3/10"),
        by = "day",
        ignore = as.Date("2000/1/25"),
        value = TRUE)
}
