% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{first_char}
\alias{first_char}
\title{Extract the first character from a string}
\usage{
first_char(string)
}
\arguments{
\item{string}{Character vector from which the first character will be extracted}
}
\value{
Character vector with the first character from each element in the vector passed to the input "string" argument. This value will be the same length as the original vector.
}
\description{
This helper function will extract the first character from a string. The element may be a letter, number, or special character but will be coerced to a character vector in the output.
}
\examples{

first_char("purple")
first_char(c("purple","rain"))
first_char(c("nothing","compares","2u"))

}
