% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{gg_zoom}
\alias{gg_zoom}
\title{Zoom in on a ggplot2 plot}
\usage{
gg_zoom(
  .plot,
  zoom_cmd,
  draw_box = TRUE,
  box_nudge = 1,
  to_label = FALSE,
  label
)
}
\arguments{
\item{.plot}{\code{ggplot2} plot object that will serve as the original plot to be "zoomed"}

\item{zoom_cmd}{Expression to return a logical value that, if \code{TRUE}, will indicate data to keep in the "zoomed" version of the plot; accepts bare column names of the data and is inherited by \code{...} internally by dplyr \link[dplyr]{filter} to perform the subsetting operation}

\item{draw_box}{Boolean indicating whether or not to try to draw a rectangle around the subset being zoomed in on for the plot. Default is \code{TRUE}. Note that this argument will only work if \emph{both} the "x" and "y" axes are on continuous scales.}

\item{box_nudge}{Coordinates to offset box if \code{draw_box = TRUE}; nudge applied to min and max of x and y coordinates for the box drawn; default is \code{1}}

\item{to_label}{Boolean indicating whether or not the plot should try to label points. Default is \code{FALSE}.}

\item{label}{Bare name of the column used to label the data points if \code{to_label = TRUE}}
}
\value{
A \link[patchwork]{patchwork} object with the \code{ggplot2} plot and the "zoomed" version side-by-side. The returned object can be further customized using functions from \code{patchwork} including \link[patchwork]{plot_layout}.
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}]

This function takes an input \code{ggplot2} object and will situate the original side-by-side with the same style plot with a user-defined subset of observations. The effect is that the original plot will be "zoomed" in on specific points.
}
\details{
\strong{NOTE}: The "zoomed" in plot will be autoscaled based on the values returned, so it may not match the scale of the original. Interpret with caution.
}
