% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{mav}
\alias{mav}
\title{Calculate moving average}
\usage{
mav(x, window, align = "center")
}
\arguments{
\item{x}{Numeric vector or time series}

\item{window}{Numeric vector of length 1 indicating the size of window for the moving average}

\item{align}{Character vector specifying the alignment of the window; allowed values include \code{"left"} or \code{"center"}; default is \code{"center"}}
}
\value{
Vector of the same length as \code{x} with the moving average at each index. Note \code{window} - 1 of the values will be \code{NA}, and depending on the \code{align} parameter these missing values will appear either at the beginning, end, or both sides of the vector.
}
\description{
This function calculates the moving average of a time series or numeric vector. The rolling calculation can be parameterized by the size of the window with the argument \code{window} and alignment of the window (centered or left-aligned) with the \code{align} parameter.
}
\details{
The functionality here is a wrapper for \link[stats]{filter}. The \code{align} argument corresponds to \code{sides} in \code{stats::filter}, with \code{align = "left"} triggering \code{sides = 1} and \code{align = "center"} triggering \code{sides = 2}. The function uses the default convolution filter from \code{stats::filter}.
}
\examples{

x <- rpois(100, lambda = 3)
mav(x, window = 7, align = "center")
mav(x, window = 7, align = "left")
}
