% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{medf}
\alias{medf}
\title{Format median}
\usage{
medf(x, digits = 2, na.rm = TRUE, ...)
}
\arguments{
\item{x}{Numeric vector to summarize}

\item{digits}{The number of decimal places to use; default is \code{2}}

\item{na.rm}{Boolean indicating whether or not \code{NA} and \code{NaN} values should be dropped; inherited by \link[stats]{fivenum}; default is \code{TRUE}}

\item{...}{Additional arguments passed to \link[base]{formatC}}
}
\value{
Character vector of length 1 with formatted median followed by 25th,75th percentiles in parentheses:\code{"median (25th,75th)"}. The formatting can be further customized using arguments to \code{base::formatC}.
}
\description{
The purpose of this function is to calculate the median, 25th and 75th percentiles of a numeric vector and format output as a string.
}
\details{
Note that the values for median, 25th, and 75th percentiles are calculated using \link[stats]{fivenum}.
}
\examples{

x <- sample(1:9, size = 100, replace = TRUE, prob = seq(0.9,0.1,-0.1))
medf(x)

}
