% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{summary_se}
\alias{summary_se}
\title{Calculate mean and standard error}
\usage{
summary_se(.data, measure_var, ..., .ci = 0.95, na.rm = TRUE)
}
\arguments{
\item{.data}{A \code{data.frame} or \code{data.frame} extension (e.g. \code{tibble})}

\item{measure_var}{Bare column name for variable to summarize}

\item{...}{Optional bare column names for grouping variables}

\item{.ci}{Width of confidence interval; default is \code{0.95}}

\item{na.rm}{Boolean indicating whether or not to use the \code{na.rm} argument in \code{mean()} and \code{sd()} functions; default is \code{TRUE}}
}
\value{
Tibble with columns for grouping variables (if any) and summarized values for mean ("mean"), standard error ("se"), and confidence interval ("ci").
}
\description{
This function will take an input \code{data.frame} and summarize a specified numeric column. The summary will include mean, standard error, and a confidence interval corresponding to the width specified (see ".ci" argument). The function leverages tidy eval internally and can accommodate bare column names.
}
\details{
The inspiration for this function was Winston Chang's \code{summarySE()} function from "The Cookbook for R". As mentioned in that text, this method of computing standard error is only appropriate for \emph{between} group variation. If there is \emph{within} subject variation (repeated measures) use another method (see references below).
}
\examples{
summary_se(.data = ToothGrowth, measure_var = len, supp, dose)
}
\references{
http://www.cookbook-r.com/Manipulating_data/Summarizing_data/

http://www.cookbook-r.com/Graphs/Plotting_means_and_error_bars_(ggplot2)/#error-bars-for-within-subjects-variables
}
