% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{theme_mathbook}
\alias{theme_mathbook}
\title{Custom "math book" \code{ggplot2} theme}
\usage{
theme_mathbook()
}
\value{
Returns a \link[ggplot2]{theme} object with custom theme settings for \code{theme_mathbook}.
}
\description{
This is a custom \code{ggplot2} theme function that inverts the default white background / black text and customizes legend placement.
}
\details{
The theme returned from this function uses:
\itemize{
\item Black panel, plot, and legend backgrounds
\item White text
\item Blank legend title
\item Legend position on bottom
}
}
